+++ src/main/java/com/proxiad/hangmangame/logic/game/GameSessionService.java
@@ -1,7 +1,7 @@
 package com.proxiad.hangmangame.logic.game;
 
 import javax.servlet.ServletException;
-import com.proxiad.hangmangame.model.GameSession;
+import com.proxiad.hangmangame.model.game.GameSession;
 
 public interface GameSessionService {
