+++ src/main/java/com/proxiad/hangmangame/logic/statistic/GameStatisticService.java
@@ -1,8 +1,8 @@
 package com.proxiad.hangmangame.logic.statistic;
 
 import java.util.Optional;
-import com.proxiad.hangmangame.model.GameSession;
-import com.proxiad.hangmangame.model.GameStatistic;
+import com.proxiad.hangmangame.model.game.GameSession;
+import com.proxiad.hangmangame.model.statistic.GameStatistic;
 
 public interface GameStatisticService {
