+++ src/main/java/com/proxiad/hangmangame/logic/statistic/GameStatisticServiceImpl.java
@@ -3,8 +3,9 @@ package com.proxiad.hangmangame.logic.statistic;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.proxiad.hangmangame.model.GameSession;
-import com.proxiad.hangmangame.model.GameStatistic;
+import com.proxiad.hangmangame.model.game.GameSession;
+import com.proxiad.hangmangame.model.statistic.GameStatistic;
+import com.proxiad.hangmangame.model.statistic.GameStatisticRepository;
 
 @Service
 public class GameStatisticServiceImpl implements GameStatisticService {
