--- src/main/java/com/proxiad/hangmangame/model/GameResult.java
@@ -1,6 +1,0 @@
-package com.proxiad.hangmangame.model;
-
-public enum GameResult {
-  WIN,
-  LOSS
-}
