+++ src/main/java/com/proxiad/hangmangame/model/game/GameSessionDaoImpl.java
@@ -1,4 +1,4 @@
-package com.proxiad.hangmangame.logic.game;
+package com.proxiad.hangmangame.model.game;
 
 import java.util.Optional;
 import javax.persistence.EntityManager;
@@ -10,7 +10,6 @@ import javax.persistence.criteria.CriteriaDelete;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import org.springframework.stereotype.Repository;
-import com.proxiad.hangmangame.model.GameSession;
 import com.proxiad.hangmangame.model.GameSession_;
 
 @Repository
