+++ src/main/java/com/proxiad/hangmangame/model/ranking/RankingRepository.java
@@ -1,10 +1,9 @@
-package com.proxiad.hangmangame.logic.ranking;
+package com.proxiad.hangmangame.model.ranking;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.proxiad.hangmangame.model.GameRanking;
 
 @Repository
 public interface RankingRepository extends JpaRepository<GameRanking, String> {
