+++ src/main/java/com/proxiad/hangmangame/model/statistic/GameResult.java
@@ -1,0 +1,6 @@
+package com.proxiad.hangmangame.model.statistic;
+
+public enum GameResult {
+  WIN,
+  LOSS
+}
