+++ src/main/java/com/proxiad/hangmangame/model/statistic/GameStatisticRepository.java
@@ -1,11 +1,10 @@
-package com.proxiad.hangmangame.logic.statistic;
+package com.proxiad.hangmangame.model.statistic;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
-import com.proxiad.hangmangame.model.GameSession;
-import com.proxiad.hangmangame.model.GameStatistic;
+import com.proxiad.hangmangame.model.game.GameSession;
 
 @Repository
 public interface GameStatisticRepository
