+++ src/main/java/com/proxiad/hangmangame/model/word/HangmanWordRepository.java
@@ -1,8 +1,7 @@
-package com.proxiad.hangmangame.logic.word;
+package com.proxiad.hangmangame.model.word;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.proxiad.hangmangame.model.HangmanWord;
 
 @Repository
 public interface HangmanWordRepository extends JpaRepository<HangmanWord, Integer> {
