+++ src/test/java/com/proxiad/hangmangame/logic/game/GameSessionServiceImplTest.java
@@ -12,9 +12,10 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import com.proxiad.hangmangame.logic.word.HangmanWordRepository;
-import com.proxiad.hangmangame.model.GameSession;
-import com.proxiad.hangmangame.model.HangmanWord;
+import com.proxiad.hangmangame.model.game.GameSession;
+import com.proxiad.hangmangame.model.game.GameSessionDao;
+import com.proxiad.hangmangame.model.word.HangmanWord;
+import com.proxiad.hangmangame.model.word.HangmanWordRepository;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
